refactor: pages and properties
