+++ src/libs/notion/properties/properties.ts
@@ -1,10 +1,7 @@
-import { callAPIWithBackOff } from "../../../utils/api.js";
-import { notion } from "../../auth.js";
+import { callAPIWithBackOff } from "../../utils/api.js";
+import { notion } from "../auth.js";
 
-import { convertResponseToPropertyValue } from "./convertResponseToPropertyValue.js";
-
-import type { PropertyValue } from "../../../../types/notion/pages/properties/properties.js";
-import type { Result } from "../../../../types/utils.js";
+import type { Result } from "../../../types/utils.js";
 import type {
   GetPagePropertyParameters,
   GetPagePropertyResponse,
@@ -37,18 +34,4 @@ export const retrievePageProperty = async (
   }
 
   return result;
-};
-
-export const fetchPageProperty = async (
-  args: GetPagePropertyParameters,
-): Promise<Result<PropertyValue>> => {
-  const { ok, data } = await retrievePageProperty(args);
-
-  if (!ok) {
-    return { ok, data };
-  }
-
-  const pagePropertyObject = await convertResponseToPropertyValue(data);
-
-  return { ok: true, data: pagePropertyObject };
 };
